serch bar and exeptions

+++ src/main/java/com/juannnnn/product_management/service/ProductService.java
@@ -1,5 +1,6 @@
 package com.juannnnn.product_management.service;
 
+import com.juannnnn.product_management.exceptions.ResourceNotFoundException;
 import com.juannnnn.product_management.model.Product;
 import com.juannnnn.product_management.repository.ProductRepository;
 import org.springframework.stereotype.Service;
@@ -15,7 +16,12 @@ public class ProductService {
         this.productRepository = productRepository;
     }
 
-    public List<Product> lisAll() {
+    public List<Product> listAll(String keyWord) {
+
+        if(keyWord != null){
+            return productRepository.findAll(keyWord);
+        }
+
         return productRepository.findAll();
     }
 
@@ -24,10 +30,13 @@ public class ProductService {
     }
 
     public Product getProductById(Long id) {
-        return productRepository.findById(id).get();
+        return productRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Product not found"));
     }
 
     public void deleteProduct(Long id) {
+        if(!productRepository.existsById(id)){
+            throw new ResourceNotFoundException("Product not found");
+        }
         productRepository.deleteById(id);
     }
